Refactor Settings, add donation features setting
